Rename task_definition.py to actions.py
Rename task_definition module to actions.py and update all imports across account and task modules to use anvil.actions.

Remove commented-out/unused task validation helper functions from task_validation.py. This consolidates the action recording logic under a clearer module name and cleans up dead code.
