+++ src/anvil/actions.py
@@ -2,18 +2,6 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
-# from anvil.results import ExecutionStatus
-#
-#
-# @dataclass(frozen=True, slots=True)
-# class TaskExecutionResult:
-#     task_name: str
-#     status: ExecutionStatus
-#     started_at: float
-#     ended_at: float
-#     result: object | None = None
-#     error: str | None = None
-
 
 @dataclass(slots=True)
 class ActionRecorder:
