+++ src/anvil/tasks/compare_asg_to_cluster_instances.py
@@ -5,7 +5,7 @@ import logging
 import boto3
 from botocore.exceptions import ClientError
 
-from anvil.task_definition import ActionRecorder
+from anvil.actions import ActionRecorder
 
 __LOGGER__ = logging.getLogger(__name__)
