+++ src/anvil/tasks/noop.py
@@ -10,7 +10,7 @@ A noop task is useful for:
 
 import logging
 
-from anvil.task_definition import ActionRecorder
+from anvil.actions import ActionRecorder
 
 __LOGGER__ = logging.getLogger(__name__)
