+++ src/anvil/tasks/remove_iam_user.py
@@ -4,7 +4,7 @@ import logging
 
 from botocore.exceptions import ClientError
 
-from anvil.task_definition import ActionRecorder
+from anvil.actions import ActionRecorder
 
 __LOGGER__ = logging.getLogger(__name__)
